+++ src/main/java/com/channelengine/merchant/apiclient/model/MerchantProductRequest.java
@@ -24,6 +24,7 @@ import com.google.gson.stream.JsonWriter;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -66,13 +67,13 @@ public class MerchantProductRequest {
   private Integer stock = null;
 
   @SerializedName("Price")
-  private Double price = null;
+  private BigDecimal price = null;
 
   @SerializedName("MSRP")
-  private Double MSRP = null;
+  private BigDecimal MSRP = null;
 
   @SerializedName("PurchasePrice")
-  private Double purchasePrice = null;
+  private BigDecimal purchasePrice = null;
 
   /**
    * The type of VAT which applies to this product.  See: http://ec.europa.eu/taxation_customs/taxation/vat/topics/rates_en.htm
@@ -127,7 +128,7 @@ public class MerchantProductRequest {
   private VatRateTypeEnum vatRateType = null;
 
   @SerializedName("ShippingCost")
-  private Double shippingCost = null;
+  private BigDecimal shippingCost = null;
 
   @SerializedName("ShippingTime")
   private String shippingTime = null;
@@ -369,7 +370,7 @@ public class MerchantProductRequest {
     this.stock = stock;
   }
 
-  public MerchantProductRequest price(Double price) {
+  public MerchantProductRequest price(BigDecimal price) {
     this.price = price;
     return this;
   }
@@ -379,15 +380,15 @@ public class MerchantProductRequest {
    * @return price
   **/
   @ApiModelProperty(value = "Price, including VAT.")
-  public Double getPrice() {
+  public BigDecimal getPrice() {
     return price;
   }
 
-  public void setPrice(Double price) {
+  public void setPrice(BigDecimal price) {
     this.price = price;
   }
 
-  public MerchantProductRequest MSRP(Double MSRP) {
+  public MerchantProductRequest MSRP(BigDecimal MSRP) {
     this.MSRP = MSRP;
     return this;
   }
@@ -397,15 +398,15 @@ public class MerchantProductRequest {
    * @return MSRP
   **/
   @ApiModelProperty(value = "Manufacturer's suggested retail price")
-  public Double getMSRP() {
+  public BigDecimal getMSRP() {
     return MSRP;
   }
 
-  public void setMSRP(Double MSRP) {
+  public void setMSRP(BigDecimal MSRP) {
     this.MSRP = MSRP;
   }
 
-  public MerchantProductRequest purchasePrice(Double purchasePrice) {
+  public MerchantProductRequest purchasePrice(BigDecimal purchasePrice) {
     this.purchasePrice = purchasePrice;
     return this;
   }
@@ -415,11 +416,11 @@ public class MerchantProductRequest {
    * @return purchasePrice
   **/
   @ApiModelProperty(value = "Optional. The purchase price of the product. Useful for repricing.")
-  public Double getPurchasePrice() {
+  public BigDecimal getPurchasePrice() {
     return purchasePrice;
   }
 
-  public void setPurchasePrice(Double purchasePrice) {
+  public void setPurchasePrice(BigDecimal purchasePrice) {
     this.purchasePrice = purchasePrice;
   }
 
@@ -441,7 +442,7 @@ public class MerchantProductRequest {
     this.vatRateType = vatRateType;
   }
 
-  public MerchantProductRequest shippingCost(Double shippingCost) {
+  public MerchantProductRequest shippingCost(BigDecimal shippingCost) {
     this.shippingCost = shippingCost;
     return this;
   }
@@ -451,11 +452,11 @@ public class MerchantProductRequest {
    * @return shippingCost
   **/
   @ApiModelProperty(value = "Shipping cost of the product.")
-  public Double getShippingCost() {
+  public BigDecimal getShippingCost() {
     return shippingCost;
   }
 
-  public void setShippingCost(Double shippingCost) {
+  public void setShippingCost(BigDecimal shippingCost) {
     this.shippingCost = shippingCost;
   }
